refactor: confirm modal
