+++ web/src/lib/modals/ConfirmModal.svelte
